# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Bingo Builder API
  description: >-
    RESTful API for creating, managing, and playing multiplayer bingo cards.


    **Security Features:**

    - CORS protection with origin restriction

    - Multi-tier rate limiting (API, write, peer, list operations)

    - Request validation via OpenAPI specification (auto-generated from Zod
    schemas)

    - Request body size limits (1MB)


    **Note:** This OpenAPI specification is automatically generated from Zod
    schemas.

    Do not edit this file manually. Instead, update the schemas in
    `backend/schemas/` and regenerate.
  version: 1.0.0
  contact:
    name: Bingo Builder
servers:
  - url: http://localhost:3001
    description: Development server
  - url: http://localhost:3002
    description: Test server
...
components:
  schemas:
...
    PeerInfo:
      type: object
      properties:
        peerId:
          type: string
          format: uuid
          description: Unique peer identifier
        playerName:
          type: string
          description: Player display name
        checkedCount:
          type: integer
          minimum: 0
          description: Number of tiles checked
      required:
        - peerId
        - playerName
        - checkedCount
    PeerRegistration:
      type: object
      properties:
        peerId:
          type: string
          format: uuid
          description: Unique peer identifier (client-generated UUID)
          example: 550e8400-e29b-41d4-a716-446655440000
        playerName:
          type: string
          minLength: 1
          maxLength: 10
          description: Player display name
          example: Alice
        checkedCount:
          type: integer
          minimum: 0
          default: 0
          description: Number of tiles checked by this player
          example: 0
      required:
        - peerId
        - playerName
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: Card not found
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
              errorCode:
                type: string
            required:
              - message
          description: Detailed validation errors (for 400 responses)
      required:
        - message
...
    PeerSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
        activePeerCount:
          type: integer
          description: Number of active peers for this card
      required:
        - success
    PeerListResponse:
      type: object
      properties:
        peers:
          type: array
          items:
            $ref: "#/components/schemas/PeerInfo"
          description: List of active peers
      required:
        - peers
    PeerHeartbeat:
      type: object
      properties:
        checkedCount:
          type: integer
          minimum: 0
          description: Updated checked count
...
paths:
...
  /api/peers/{cardId}/register:
    post:
      summary: Register or update a peer
      description: Register a player as an active peer for a card or update their
        information
      tags:
        - peers
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
            description: MongoDB ObjectId of the card
            example: 507f1f77bcf86cd799439011
          required: true
          description: MongoDB ObjectId of the card
          name: cardId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeerRegistration"
      responses:
        "200":
          description: Peer registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeerSuccessResponse"
        "400":
          description: Invalid request parameters or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded or max peers reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/peers/{cardId}/peers:
    get:
      summary: Get active peers
      description: Retrieve list of active peers for a card
      tags:
        - peers
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
            description: MongoDB ObjectId of the card
            example: 507f1f77bcf86cd799439011
          required: true
          description: MongoDB ObjectId of the card
          name: cardId
          in: path
        - schema:
            type: string
            format: uuid
            description: Peer ID to exclude from results (usually the requesting peer)
          required: false
          description: Peer ID to exclude from results (usually the requesting peer)
          name: excludePeerId
          in: query
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeerListResponse"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/peers/{cardId}/heartbeat/{peerId}:
    post:
      summary: Send peer heartbeat
      description: Keep peer connection alive and update checked count
      tags:
        - peers
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
            description: MongoDB ObjectId of the card
            example: 507f1f77bcf86cd799439011
          required: true
          description: MongoDB ObjectId of the card
          name: cardId
          in: path
        - schema:
            type: string
            format: uuid
            description: UUID of the peer
            example: 550e8400-e29b-41d4-a716-446655440000
          required: true
          description: UUID of the peer
          name: peerId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeerHeartbeat"
      responses:
        "200":
          description: Heartbeat successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/peers/{cardId}/unregister/{peerId}:
    delete:
      summary: Unregister a peer
      description: Remove a peer from the active peer list (when leaving)
      tags:
        - peers
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
            description: MongoDB ObjectId of the card
            example: 507f1f77bcf86cd799439011
          required: true
          description: MongoDB ObjectId of the card
          name: cardId
          in: path
        - schema:
            type: string
            format: uuid
            description: UUID of the peer
            example: 550e8400-e29b-41d4-a716-446655440000
          required: true
          description: UUID of the peer
          name: peerId
          in: path
      responses:
        "200":
          description: Peer unregistered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...